Fixed an issue in player settings

+++ mcnative-runtime-common/src/main/java/org/mcnative/runtime/common/player/data/DefaultPlayerDataProvider.java
@@ -174,7 +174,7 @@ public class DefaultPlayerDataProvider implements PlayerDataProvider {
 
     @Override
     public void deleteSetting(PlayerSetting setting) {
-        Validate.notNull(setting);
+        if(setting == null) return;
         playerSettingsStorage.delete()
                 .where("Id",setting.getId())
                 .execute();
